Reorganize the KafkaStreams beans initialization
This will avoid race conditions and make sure everything is properly
initialized before consuming the beans.
