+++ extensions/kafka-streams/runtime/src/main/java/io/quarkus/kafka/streams/runtime/KafkaStreamsRecorder.java
@@ -1,12 +1,10 @@
 package io.quarkus.kafka.streams.runtime;
 
 import java.util.Properties;
+import java.util.function.Supplier;
 
 import org.rocksdb.RocksDB;
 
-import io.quarkus.arc.Arc;
-import io.quarkus.arc.runtime.BeanContainer;
-import io.quarkus.arc.runtime.BeanContainerListener;
 import io.quarkus.runtime.annotations.Recorder;
 
 @Recorder
@@ -16,17 +14,11 @@ public class KafkaStreamsRecorder {
         RocksDB.loadLibrary();
     }
 
-    public void configureRuntimeProperties(KafkaStreamsRuntimeConfig runtimeConfig) {
-        Arc.container().instance(KafkaStreamsTopologyManager.class).get().setRuntimeConfig(runtimeConfig);
-    }
-
-    public BeanContainerListener configure(Properties properties) {
-        return new BeanContainerListener() {
-
+    public Supplier<KafkaStreamsSupport> kafkaStreamsSupportSupplier(Properties properties) {
+        return new Supplier<KafkaStreamsSupport>() {
             @Override
-            public void created(BeanContainer container) {
-                KafkaStreamsTopologyManager instance = container.instance(KafkaStreamsTopologyManager.class);
-                instance.configure(properties);
+            public KafkaStreamsSupport get() {
+                return new KafkaStreamsSupport(properties);
             }
         };
     }
