+++ extensions/kafka-streams/runtime/src/main/java/io/quarkus/kafka/streams/runtime/KafkaStreamsSupport.java
@@ -1,0 +1,16 @@
+package io.quarkus.kafka.streams.runtime;
+
+import java.util.Properties;
+
+public class KafkaStreamsSupport {
+
+    private final Properties properties;
+
+    public KafkaStreamsSupport(Properties properties) {
+        this.properties = properties;
+    }
+
+    public Properties getProperties() {
+        return properties;
+    }
+}
