+++ extensions/kafka-streams/runtime/src/main/java/io/quarkus/kafka/streams/runtime/health/KafkaStreamsStateHealthCheck.java
@@ -9,20 +9,18 @@ import org.eclipse.microprofile.health.HealthCheckResponse;
 import org.eclipse.microprofile.health.HealthCheckResponseBuilder;
 import org.eclipse.microprofile.health.Liveness;
 
-import io.quarkus.kafka.streams.runtime.KafkaStreamsTopologyManager;
-
 @Liveness
 @ApplicationScoped
 public class KafkaStreamsStateHealthCheck implements HealthCheck {
 
     @Inject
-    private KafkaStreamsTopologyManager manager;
+    private KafkaStreams kafkaStreams;
 
     @Override
     public HealthCheckResponse call() {
         HealthCheckResponseBuilder responseBuilder = HealthCheckResponse.named("Kafka Streams state health check");
         try {
-            KafkaStreams.State state = manager.getStreams().state();
+            KafkaStreams.State state = kafkaStreams.state();
             responseBuilder.state(state.isRunningOrRebalancing())
                     .withData("state", state.name());
         } catch (Exception e) {
